split test files out
