+++ package.json
@@ -17,7 +17,7 @@
     "tape": "^4.2.2"
   },
   "scripts": {
-    "test": "node test/index.js",
+    "test": "tape 'test/*.js'",
     "prepublish": "npm ls && npm test"
   },
   "author": "Dominic Tarr <dominic.tarr@gmail.com> (http://dominictarr.com)",
